fix: Renames example_class.yaml to example.yaml
