+++ templates/serverpod_templates/projectname_client/lib/src/protocol/example.dart
