+++ templates/serverpod_templates/projectname_client/lib/src/protocol/protocol.dart
@@ -8,8 +8,8 @@
 library protocol; // ignore_for_file: no_leading_underscores_for_library_prefixes
 
 import 'package:serverpod_client/serverpod_client.dart' as _i1;
-import 'example_class.dart' as _i2;
-export 'example_class.dart';
+import 'example.dart' as _i2;
+export 'example.dart';
 export 'client.dart'; // ignore_for_file: equal_keys_in_map
 
 class Protocol extends _i1.SerializationManager {
