+++ templates/serverpod_templates/projectname_server/lib/src/protocol/example.yaml
